+++ package.json
@@ -58,6 +58,7 @@
     "react-dom": "^16.10.1",
     "react-lazy-load": "^3.0.13",
     "react-router-dom": "^5.1.2",
+    "reselect": "^4.0.0",
     "resolve": "1.12.0",
     "resolve-url-loader": "3.1.0",
     "sass-loader": "7.2.0",
