+++ src/containers/index.js
@@ -1,6 +1,7 @@
 import RouteForm from "./route-form";
 import RouteMap from "./route-map";
+import ResultsSidebar from "./results-sidebar";
 import Results from "./results";
 import AddStopForm from "./add-stop-form";
 
-export { AddStopForm, RouteForm, RouteMap, Results };
+export { AddStopForm, RouteForm, RouteMap, ResultsSidebar, Results };
